Created some of the templates for the website.
base.html is a simple base template, that all other pages inherit from.

Navbar is the navbar that shows up on every page in the site.

Flash messages is used to display flashed errors to the user.

blog_post.html is a generic page template, Simply including the Navbar and any content blocks associated with the page.

Signed-off-by: EdMan1022 <EdMan1022@gmail.com>